fix(review): extract duplicate _has_pytestmark_cli to shared _split_helpers module
_has_pytestmark_cli was defined identically in test_update_checks_split.py
and test_doctor_split.py. Extracted to tests/cli/_split_helpers.py following
the established pattern of _fleet_helpers.py and _update_checks_helpers.py;
both split-guard files now import from the shared module.
