+++ tests/cli/test_update_checks_split.py
@@ -2,10 +2,11 @@
 
 from __future__ import annotations
 
-import ast
 from pathlib import Path
 
 import pytest
+
+from ._split_helpers import _has_pytestmark_cli
 
 pytestmark = [pytest.mark.layer("cli"), pytest.mark.small]
 
@@ -40,17 +41,6 @@ def test_order_importable_from_submodule():
 
 
 _CLI_TESTS = Path(__file__).parent
-
-
-def _has_pytestmark_cli(path: Path) -> bool:
-    tree = ast.parse(path.read_text())
-    for node in ast.walk(tree):
-        if isinstance(node, ast.Assign):
-            for target in node.targets:
-                if isinstance(target, ast.Name) and target.id == "pytestmark":
-                    src = ast.unparse(node.value)
-                    return 'layer("cli")' in src or "layer('cli')" in src
-    return False
 
 
 def test_update_checks_lifecycle_file_exists():
